+++ src/integration-test/java/com/synopsys/defensics/jenkins/test/RunPipelineIT.java
@@ -24,7 +24,6 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
-import static org.hamcrest.Matchers.nullValue;
 
 import com.synopsys.defensics.apiserver.model.RunState;
 import com.synopsys.defensics.apiserver.model.RunVerdict;
@@ -188,11 +187,10 @@ public class RunPipelineIT {
     dumpRunLog(run);
 
     assertThat(run.getResult(), is(equalTo(Result.ABORTED)));
-    assertThat(run.getActions(HtmlReportAction.class).size(), is(equalTo(0)));
+    assertThat(run.getActions(HtmlReportAction.class).size(), is(equalTo(1)));
     assertThat(run.getActions(HTMLAction.class).size(), is(equalTo(0)));
-    assertThat(project.getActions(HtmlReportAction.class).size(), is(equalTo(0)));
-    assertThat(project.getAction(HtmlReportAction.class),
-        is(nullValue()));
+    assertThat(project.getActions(HtmlReportAction.class).size(), is(equalTo(1)));
+    assertThat(project.getAction(HtmlReportAction.class), is(notNullValue()));
   }
 
   @Test